requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.3.1
google-generativeai==0.3.2
rank-bm25==0.2.2
openpyxl==3.1.2
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3

